requirements0211.txt
about-time==4.2.1
alive-progress==3.1.5
annotated-types==0.7.0
autograd==1.6.2
certifi==2024.6.2
charset-normalizer==3.3.2
clarabel==0.9.0
click==8.1.7
cloudpickle==3.0.0
cma==3.2.2
contourpy==1.2.1
cvxpy==1.5.2
cycler==0.12.1
decorator==4.4.2
Deprecated==1.2.14
dill==0.3.8
docker-pycreds==0.4.0
ecos==2.0.14
Farama-Notifications==0.0.4
filelock==3.15.4
fire==0.6.0
fonttools==4.53.0
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
grapheme==0.6.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mo-gymnasium==1.1.0
morl-baselines==1.0.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
osqp==0.6.7.post0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
proglog==0.1.10
protobuf==5.27.2
psutil==6.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.0
Pygments==2.18.0
pymoo==0.6.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qdldl==0.1.7.post4
requests==2.32.3
rich==13.9.3
scalene==1.5.49
scikit-learn==1.5.0
scipy==1.14.0
scs==3.2.5
seaborn==0.13.2
sentry-sdk==2.7.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12.1
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.19.4
wrapt==1.16.0
